+++ Actions/AL-Go-Helper.ps1
@@ -652,7 +652,7 @@ function ReadSettings {
         "incrementalBuilds"                             = [ordered]@{
             "enable"                                    = $false
             "retentionDays"                             = 30
-            "mode"                                      = "modifiedApps" # modifiedProjects, modifiedApps, modifiedAppsAndDependingApps
+            "mode"                                      = "modifiedApps" # modifiedProjects, modifiedApps
         }
         "microsoftTelemetryConnectionString"            = "InstrumentationKey=cd2cc63e-0f37-4968-b99a-532411a314b8;IngestionEndpoint=https://northeurope-2.in.applicationinsights.azure.com/"
         "partnerTelemetryConnectionString"              = ""