+++ Actions/DetermineProjectsToBuild/DetermineProjectsToBuild.Action.ps1
@@ -36,6 +36,7 @@ elseif ($ghEvent.PSObject.Properties.name -eq 'workflow_dispatch') {
 else {
     # Push
     $buildAllProjects = !$settings.incrementalBuilds.enable
+
     $branch = $env:GITHUB_REF_NAME
     Write-Host "Push on $branch"
 }